edit2

+++ Sensor.h
@@ -14,7 +14,7 @@ class Sensor{
 	double data[3];
 public:
 	Sensor(int n_pin1,int n_pin2);
-	float getdistance(int trigger,int echo);
+	double getdistance(int trigger,int echo);
 private:
 	double median(Iterator begin, Iterator end);
 };
